Moving aws_manager to /src
